searching linear

+++ 0x1E-search_algorithms/0-linear.c
@@ -12,7 +12,7 @@ int linear_search(int *array, size_t size, int value)
 
 	if (size <= 0 || array == NULL)
 		return (-1);
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 	{
 		printf("Value checked array[%i] = [%i]\n",
 			   i, array[i]);
